backend/pyproject.toml
[project]
name = "gini-backend"
version = "0.1.0"
description = "Gini AI Assistant Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0.2",
    "openai>=1.58.0",
    "anthropic>=0.40.0",
    "cryptography>=44.0.0",
    "pgvector>=0.3.0",
]

[tool.setuptools.packages.find]
include = ["app*", "scripts*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
